refactor: refactor all services in their correct package 'interfaces'
