+++ back/healthapp/src/main/java/com/healthcare/domain/service/interfaces/IAppointmentService.java
@@ -1,4 +1,4 @@
-package com.healthcare.domain.service;
+package com.healthcare.domain.service.interfaces;
 
 import jakarta.mail.MessagingException;
 import org.springframework.http.ResponseEntity;
