+++ back/healthapp/src/main/java/com/healthcare/domain/service/interfaces/IMedicService.java
@@ -1,4 +1,4 @@
-package com.healthcare.domain.service;
+package com.healthcare.domain.service.interfaces;
 
 import org.springframework.http.ResponseEntity;
